requirements-dev.txt
# ============================================================================
# Development Dependencies for Coal LIMS
# ============================================================================
# Энэ файлыг development орчинд л суулгана:
# pip install -r requirements-dev.txt

# Include production dependencies
-r requirements.txt

# ============================================================================
# TESTING
# ============================================================================

# Pytest framework
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Coverage reporting
coverage[toml]==7.3.2

# Fixtures and factories
factory-boy==3.3.0
Faker==20.1.0

# ============================================================================
# CODE QUALITY
# ============================================================================

# Code formatting
black==23.11.0
isort==5.12.0

# Linting
flake8==6.1.0
flake8-docstrings==1.7.0
flake8-bugbear==23.11.28
flake8-comprehensions==3.14.0
pylint==3.0.2

# Type checking
mypy==1.7.1
types-python-dateutil==2.8.19.14
types-pytz==2023.3.1.1

# Security scanning
bandit==1.7.5
safety==2.3.5

# ============================================================================
# DEVELOPMENT TOOLS
# ============================================================================

# Pre-commit hooks
pre-commit==3.5.0

# Debugging
ipdb==0.13.13
ipython==8.17.2

# Performance profiling
py-spy==0.3.14

# HTTP testing
requests-mock==1.11.0

# ============================================================================
# DOCUMENTATION
# ============================================================================

# Sphinx documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# API documentation
flasgger==0.9.7.1

# ============================================================================
# DATABASE TOOLS
# ============================================================================

# Database inspection
sqlalchemy-utils==0.41.1

# Migration helpers
alembic-utils==0.8.1

# ============================================================================
# OPTIONAL: ADVANCED TESTING
# ============================================================================

# Load testing (ашиглах бол uncomment хий)
locust==2.17.0

# Selenium browser testing (ашиглах бол uncomment хий)
# selenium==4.15.2
# webdriver-manager==4.0.1

# ============================================================================
# OPTIONAL: CI/CD TOOLS
# ============================================================================

# Code coverage upload
# codecov==2.1.13

# Linting automation
# pre-commit-hooks==4.5.0
